# frontend/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KBV2 Knowledge Base API",
    "description": "High-fidelity information extraction and graph visualization API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "documents",
          "documents"
        ],
        "summary": "Get Document",
        "description": "GET /api/v1/documents/{document_id}\n\nGet document metadata and extraction statistics.\nAIP-131 Get method for retrieving resource by ID.",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Document UUID",
              "title": "Document Id"
            },
            "description": "Document UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DocumentResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/content": {
      "get": {
        "tags": [
          "documents",
          "documents"
        ],
        "summary": "Get Document Content",
        "description": "GET /api/v1/documents/{document_id}/content\n\nGet document content with all chunks concatenated in order.\nSupports content extraction state and metadata.",
        "operationId": "get_document_content_api_v1_documents__document_id__content_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Document UUID",
              "title": "Document Id"
            },
            "description": "Document UUID"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(text|json)$",
              "description": "Response format",
              "default": "text",
              "title": "Format"
            },
            "description": "Response format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DocumentContentResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/spans": {
      "get": {
        "tags": [
          "documents",
          "documents"
        ],
        "summary": "Get Document Spans",
        "description": "GET /api/v1/documents/{document_id}/spans\n\nGet all text spans with entity references for exact text highlighting.\nCalculates absolute character offsets using the OffsetCalculationService.\n\nCritical for \"Click entity -> See text evidence\" functionality.\nImplements W3C Web Annotation Data Model compliance.",
        "operationId": "get_document_spans_api_v1_documents__document_id__spans_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Document UUID",
              "title": "Document Id"
            },
            "description": "Document UUID"
          },
          {
            "name": "confidence_threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Minimum confidence threshold",
              "default": 0.0,
              "title": "Confidence Threshold"
            },
            "description": "Minimum confidence threshold"
          },
          {
            "name": "entity_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity types",
              "title": "Entity Types"
            },
            "description": "Filter by entity types"
          },
          {
            "name": "verified_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return spans with grounding quotes",
              "default": false,
              "title": "Verified Only"
            },
            "description": "Only return spans with grounding quotes"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DocumentSpansResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/entities": {
      "get": {
        "tags": [
          "documents",
          "documents"
        ],
        "summary": "Get Document Entities",
        "description": "GET /api/v1/documents/{document_id}/entities\n\nGet all entities extracted from document with their locations and confidence scores.\nGroups multiple occurrences of the same entity.",
        "operationId": "get_document_entities_api_v1_documents__document_id__entities_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Document UUID",
              "title": "Document Id"
            },
            "description": "Document UUID"
          },
          {
            "name": "entity_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity types",
              "title": "Entity Types"
            },
            "description": "Filter by entity types"
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Minimum confidence",
              "default": 0.5,
              "title": "Min Confidence"
            },
            "description": "Minimum confidence"
          },
          {
            "name": "include_spans",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include text locations",
              "default": true,
              "title": "Include Spans"
            },
            "description": "Include text locations"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DocumentEntitiesResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/:search": {
      "post": {
        "tags": [
          "documents",
          "documents"
        ],
        "summary": "Search Documents",
        "description": "POST /api/v1/documents:search\n\nHybrid search combining vector similarity (semantic) and keyword search.\nUses Reciprocal Rank Fusion for result merging with weights:\n- Vector search: 0.7 weight (semantic meaning)\n- Keyword search: 0.3 weight (exact matches)\n\nAIP-136 Custom Method for search operations.",
        "operationId": "search_documents_api_v1_documents__search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DocumentSearchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_DocumentResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if request succeeded"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload on success"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details on failure"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Additional metadata (pagination, timing, request_id)"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "APIResponse[DocumentResponse]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}